rename interface ts file to .interface.ts
