+++ core/new-gui/src/app/common/rxjs-operators.ts
@@ -8,6 +8,10 @@
 
 // Observable class extensions
 import 'rxjs/add/observable/of';
+import 'rxjs/add/observable/from';
+import 'rxjs/add/observable/fromEvent';
+import 'rxjs/add/observable/merge';
+import 'rxjs/add/observable/empty';
 
 // Observable operators
 import 'rxjs/add/operator/map';
@@ -21,8 +25,10 @@ import 'rxjs/add/operator/auditTime';
 import 'rxjs/add/operator/distinctUntilChanged';
 import 'rxjs/add/operator/startWith';
 import 'rxjs/add/operator/shareReplay';
+import 'rxjs/add/operator/share';
 import 'rxjs/add/operator/first';
 import 'rxjs/add/operator/last';
 import 'rxjs/add/operator/delay';
-import 'rxjs/add/observable/fromEvent';
+import 'rxjs/add/operator/do';
+import 'rxjs/add/operator/merge';
