+++ core/new-gui/src/app/workspace/component/operator-panel/operator-label/operator-label.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input, AfterViewInit } from '@angular/core';
 import { v4 as uuid } from 'uuid';
 
-import { OperatorSchema } from '../../../types/operator-schema';
+import { OperatorSchema } from '../../../types/operator-schema.interface';
 
 /**
  * OperatorLabelComponent is one operator box in the operator panel.
