+++ core/new-gui/src/app/workspace/service/joint-ui/joint-ui.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { OperatorMetadataService } from '../operator-metadata/operator-metadata.service';
-import { OperatorSchema } from '../../types/operator-schema';
+import { OperatorSchema } from '../../types/operator-schema.interface';
 
 import * as joint from 'jointjs';
 import { OperatorPort } from '../../types/operator-port';
@@ -66,7 +66,7 @@ export class JointUIService {
   public static readonly DEFAULT_OPERATOR_WIDTH = 140;
   public static readonly DEFAULT_OPERATOR_HEIGHT = 40;
 
-  private operators: OperatorSchema[] = [];
+  private operators: ReadonlyArray<OperatorSchema> = [];
 
 
   constructor(
