+++ core/new-gui/src/app/workspace/service/operator-metadata/operator-metadata.service.ts
@@ -6,7 +6,7 @@ import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import '../../../common/rxjs-operators';
 
 import { AppSettings } from '../../../common/app-setting';
-import { OperatorMetadata } from '../../types/operator-schema';
+import { OperatorMetadata } from '../../types/operator-schema.interface';
 
 export const MOCK_OPERATOR_METADATA_ENDPOINT = 'resources/operator-metadata';
