+++ core/new-gui/src/app/workspace/service/operator-metadata/stub-operator-metadata.service.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 
 import { getMockOperatorMetaData } from './mock-operator-metadata.data';
-import { OperatorMetadata } from '../../types/operator-schema';
+import { OperatorMetadata } from '../../types/operator-schema.interface';
 
 import '../../../common/rxjs-operators';
 import { EMPTY_OPERATOR_METADATA } from './operator-metadata.service';
